SUBAPI-7: Added MediatR Integration
