+++ Domain.Abstractions.UnitTests/Domain.Abstractions.UnitTests.csproj
@@ -7,11 +7,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
-      <PackageReference Include="NUnit" Version="3.12.0" />
-    </ItemGroup>
-
-    <ItemGroup>
       <ProjectReference Include="..\Domain.Abstractions\Domain.Abstractions.csproj" />
     </ItemGroup>
