+++ Domain.Abstractions/Domain.Abstractions.csproj
@@ -6,9 +6,4 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
-    <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.7" />
-      <PackageReference Include="MongoDB.Driver" Version="2.11.0" />
-    </ItemGroup>
-
 </Project>
