+++ Domain.Instructor/Domain.Instructor.csproj
@@ -11,4 +11,10 @@
       <ProjectReference Include="..\Domain.User\Domain.User.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+        <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="8.1.0" />
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.7" />
+        <PackageReference Include="MongoDB.Driver" Version="2.11.0" />
+    </ItemGroup>
+
 </Project>
